+++ tests/testthat/test-theme_elements_gtsummary.R
@@ -419,14 +419,14 @@ test_that("theme element add_ci.tbl_svysummary-arg:method", {
     )
 })
 
-# pkgwide-str:footnote_symbol --------------------------------------------------
-test_that("pkgwide-str:footnote_symbol", {
+# pkgwide-chr:footnote_symbol --------------------------------------------------
+test_that("pkgwide-chr:footnote_symbol", {
   skip_if_pkg_not_installed("flextable")
 
   # theme element supplies the default footnote reference symbols
   ft_calls <-
     with_gtsummary_theme(
-      x = list("pkgwide-str:footnote_symbol" = c("*", "\u2020", "\u2021")),
+      x = list("pkgwide-chr:footnote_symbol" = c("*", "\u2020", "\u2021")),
       expr =
         trial |>
           tbl_summary(by = trt, include = c(age, grade), missing = "no") |>